Programming/backend/requirements.txt
bcrypt
fastapi
mysql-connector-python
uvicorn
pydantic
pyjwt